Fix slurm management bug that was erroring out on some machines.
This fix is at the recommendation of RC. This job was being
assigned to summit nodes whose /tmp space was filled up and not
necessarily cleaned up. It's not entirely clear why this would
happen, but it was causing this job to regularly (depending on
which node it was running on and the state of that node's tmp
space) crash by running out of tmp space.  This fix tells Matlab
to use this area in scratch for tmp space instead.  Since I put
this fix in place we have had no more unexpected failures.