Set up inventory maker to work for degenerate cases.
Move the inventory maker from static to non-static so that I can
use the historicEndDt property to limit end of historic
inventory. In the event that a tile has never been inventoried
and has no actual data on the current file system, sets up the
degenerate inventory for yesterday to today so that subsequent
batch processing will have an inventory to start with.