Optimizing Step0 settings for JPL sync.
Change Step0 scheduling to be started before this processing
runs, so that the daily pipeline will run again tomorrow even if
today's Step0 fails.  Changed Step0 schedule time from 10:30 to
04:30.  JPL jobs are supposed to run at 2:30 Pacific, so this
schedules our pipeline early in the day and hopefully the
overnight queues will have emptied out.

Increase time for this job to 2.5 hours.  Normally this Step runs
in 3-4 minutes, but last week it timed out and this morning it
took over 2 hours to run.