Adding controls for IDW on RF/DV.
Added calls to spatially interpolate RF and DV layers. This eliminates the big
areas where these layers have been returning zeros (or very small values).
Currently this can be controlled with the sbatch updateSTC script, and values
for sthreshForRF are being tested and evaluated.  Eventually a default value
should be selected, and the call to updateSTC_SCAGDRFSFor should be modified so
that the default value is used and sthresh/sthreshForRF are not required inputs.
Otherwise, Step1 scripts will need to accept inputs for these and pass them
along.