Workaround for bug found on integration.
spires_albedo needs inputs to have NaNs, so used nodata values for solar zenith
and azimuth to set these to NaNs.  Found bug in solar_azimuth nodata value, it's
set to max(uint16) but should be set to max(int16).

I need to discuss the true fixes for both of these problems with Sebastien.