Fixes to get daily pipeline running.
Added version string to scagdrfs_geotiff directory, this is required
in order for scratchShuffle to work as expected.
The long string used for directory hierarchy here should be moved to
a function in ESPEnv, in order for that function to be the place to go for
anything environment-specific.

Reduced number of workers, because writeGeotiff uses up to 80GB memory.
This may be a memory leak and should probably be investigated.
Made output verbose at file level, to better track progress in output.