A few tweaks to pipeline scripts.
Step0 timed out on 2 hours this week, so extended time to 8 hours, but
it's been running in 6 minutes since then. I have a ticket in to CURC about
this.

Fixed a few things with the slurm stdout filenames.

Modified the Step3 call to write geotiffer to take advantage of monthWindow 0.